requirements.txt
pandas>=2.0.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
kafka-python>=2.0.0
numpy>=1.24.0